backend/requirements.txt
fastapi[standard]
python-multipart
numpy
opencv-python
ultralytics
Pillow 